requirements_desktop.txt
# requirements_desktop.txt
# All dependencies needed to run AND package the app as a Windows desktop exe.
# Install with:  pip install -r requirements_desktop.txt

# ── Core Django stack (already in requirements.txt) ──────────────────────
django>=4.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
whitenoise
dj-database-url

# ── ML stack ─────────────────────────────────────────────────────────────
scikit-learn
pandas
numpy
joblib

# ── Desktop wrapper ───────────────────────────────────────────────────────
# pywebview creates the native window (uses WebView2 on Windows)
pywebview>=4.4

# ── Windows WebView2 backend for pywebview ────────────────────────────────
# On Windows, pywebview uses the Microsoft Edge WebView2 runtime.
# WebView2 is pre-installed on Windows 10 (1803+) and Windows 11.
# If not present, the user can install it from:
# https://developer.microsoft.com/en-us/microsoft-edge/webview2/
# No extra pip package needed — it's part of Windows.

# ── Packaging ─────────────────────────────────────────────────────────────
pyinstaller>=6.0
